refactor(client): rename file to Imperial.ts
