+++ src/lib/index.ts
@@ -1,8 +1,8 @@
 // Export client
-export * from "./client";
+export * from "./client/Imperial";
 // Export all structures
 export * from "./classes";
 // Export all errors for the user
 export * from "./errors";
 // export Imperial as default
-export { Imperial as default } from "./client";
+export { Imperial as default } from "./client/Imperial";
