+++ src/lib/utils/decorators.ts
@@ -1,5 +1,6 @@
 import { Error as ImpError } from "../errors";
 import { Base } from "../client/Base";
+
 /**
  *  Check if the client has a token provided
  *  Method must be a promise
@@ -14,13 +15,16 @@ export const requireToken: MethodDecorator = (_target, _key, descriptor) => {
 
 	return {
 		value: function value(...args: any[]) {
+			// this should never be throw???!?!?!?!
 			if (!(this instanceof Base)) {
 				throw new Error("type check");
 			}
 
+			// reject a promise because the method is async
 			if (!this.client._token) return Promise.reject(new ImpError("NO_TOKEN"));
 
-			return originalMethod.apply(this, args);
+			// return the orignal method
+			return originalMethod.apply(this, args as string[]);
 		},
 	} as any;
 };
